treat stdout and stderr as stageable Files
in addition to letting them be strings with no staging behaviour

look at treating stdout and stderr as stageable Files not strings
(with default behaviour being no-op file staging, so preserves the
behaviour of "do nothing" - or at least that should be a goal of this
PR)

eg issue  #363

but not issue #441 - that's to do with channel usageb

TODO: could be some interesting interactions with executor std* staging?
eg. what does that look like for wq/tv? at least make a test case?

for testing there's no easy output staging provder - is there a test of
output staging using outputs= and if not, should I make one as a separate
test?

kwargs type violation: some of the code assumes that kwargs stdout will be
a string - eg in the taskrecord kwargs dict (which is generally
untyped...)