Move PolledExecutorFacade status into BlockProviderExecutor
This is intended to move all status structures into the same object.
This PR leaves accesses to those structures in the PolledExecutorFacade,
with future PRs intended to evaporate away more of the PolledExecutorFacade
and perform those accesses directly inside the BlockProviderExecutor.

This should not change behaviour, as it only moves an attribute from one
objected to its 1:1 paired object.